Fix: buyPrice Double로 변환

+++ src/main/java/com/fund/stockProject/experiment/dto/ExperimentStatusDetailResponse.java
@@ -16,8 +16,8 @@ public class ExperimentStatusDetailResponse {
     private List<TradeInfo> tradeInfos; // 거래 내역
     private Integer buyScore; // 매수 시점 점수
     private Integer currentScore; // 현재 시점 점수
-    private Integer buyPrice; // 매수 가격
-    private Integer currentPrice; // 현재 가격
+    private Double buyPrice; // 매수 가격
+    private Double currentPrice; // 현재 가격
     private LocalDateTime buyAt; // 매수일
     private COUNTRY country; // 국가
